# FrontEnd1/tools/swagger/src/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Swagger Api documentation for Gestion d'analyse medical  application",
    "version" : "0.0.1-SNAPSHOT",
    "title" : "Swagger Api Documentation for Gestion d'analyse medical application",
    "contact" : {
      "name" : "khalid Abba",
      "url" : "https://www.udemy.com/user/",
      "email" : "khalidoabba1996@gmail.com"
    }
  },
  "basePath" : "http://localhost:8080",
...
  "paths" : {
...
    "/testCovid/v1/absence/{idAbsence}" : {
      "get" : {
        "tags" : [ "testCovidv1absence" ],
        "summary" : "rechercher un absence par ID",
        "description" : "Cette methode permet de rechercher  un absence par son ID ",
        "operationId" : "findById",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "idAbsence",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "absence a ete trouver dans la BDD",
            "schema" : {
              "$ref" : "#/definitions/AbsenceDTO"
            }
          },
          "404" : {
            "description" : "Aucun absence n'existe  dans la BDD avec l'ID fourni"
          }
        }
      }
    },
...
    "/testCovid/v1/analyseMedical/filter{idAnalyseMedical}" : {
      "get" : {
        "tags" : [ "analyseMedical" ],
        "summary" : "rechercher un analyse medical par ID",
        "description" : "Cette methode permet de rechercher  d'analyse medical par son ID ",
        "operationId" : "findById",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "idAnalyseMedical",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "l'analyse medical a ete trouver dans la BDD",
            "schema" : {
              "$ref" : "#/definitions/AnalyseMedicalDTO"
            }
          },
          "404" : {
            "description" : "Aucun analyse medical  n'existe  dans la BDD avec l'ID fourni"
          }
        }
      }
    },
...
    "/testCovid/v1/article/{idArticle}" : {
      "get" : {
        "tags" : [ "article" ],
        "summary" : "rechercher un article par ID",
        "description" : "Cette methode permet de rechercher  un article par son ID ",
        "operationId" : "findById",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "idArticle",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "article a ete trouver dans la BDD",
            "schema" : {
              "$ref" : "#/definitions/ArticleDTO"
            }
          },
          "404" : {
            "description" : "Aucun article n'existe  dans la BDD avec l'ID fourni"
          }
        }
      }
    },
...
    "/testCovid/v1/bilanfinance/{bilanfinance}" : {
      "get" : {
        "tags" : [ "testCovidv1bilanfinance" ],
        "summary" : "rechercher un bilanfinance par ID",
        "description" : "Cette methode permet de rechercher  un bilanfinance par son ID ",
        "operationId" : "findById",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "idBilan",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : " bilanfinance a ete trouver dans la BDD",
            "schema" : {
              "$ref" : "#/definitions/BilanFinancierDTO"
            }
          },
          "404" : {
            "description" : "Aucun bilanfinance n'existe  dans la BDD avec l'ID fourni"
          }
        }
      }
    },
...
    "/testCovid/v1/biologiste/{idBiologiste}" : {
      "get" : {
        "tags" : [ "biologiste" ],
        "summary" : "rechercher un biologiste par ID",
        "description" : "Cette methode permet de rechercher  un biologiste par son ID ",
        "operationId" : "findById",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "idBiologiste",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "la biologiste a ete trouver dans la BDD",
            "schema" : {
              "$ref" : "#/definitions/BiologisteDTO"
            }
          },
          "404" : {
            "description" : "Aucun biologiste n'existe  dans la BDD avec l'ID fourni"
          }
        }
      }
    },
...
    "/testCovid/v1/boncommande/filter{idBonCommande}" : {
      "get" : {
        "tags" : [ "boncommande" ],
        "summary" : "rechercher boncommande par ID",
        "description" : "Cette methode permet de rechercher  boncommande par son ID ",
        "operationId" : "findById",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "idBonCommande",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "boncommandea ete trouver dans la BDD",
            "schema" : {
              "$ref" : "#/definitions/BonCommandeDTO"
            }
          },
          "404" : {
            "description" : "boncommande medical  n'existe  dans la BDD avec l'ID fourni"
          }
        }
      }
    },
...
    "/testCovid/v1/compte/{idCompte}" : {
      "get" : {
        "tags" : [ "testCovidv1compte" ],
        "summary" : "rechercher un compte par ID",
        "description" : "Cette methode permet de rechercher  un compte par son ID ",
        "operationId" : "findById",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "idCompte",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "compte a ete trouver dans la BDD",
            "schema" : {
              "$ref" : "#/definitions/CompteDTO"
            }
          },
          "404" : {
            "description" : "Aucun compte n'existe  dans la BDD avec l'ID fourni"
          }
        }
      }
    },
...
    "/testCovid/v1/conge/{idConge}" : {
      "get" : {
        "tags" : [ "testCovidv1conge" ],
        "summary" : "rechercher un conge par ID",
        "description" : "Cette methode permet de rechercher  un conge par son ID ",
        "operationId" : "findById",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "idConge",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "conge a ete trouver dans la BDD",
            "schema" : {
              "$ref" : "#/definitions/CongeDTO"
            }
          },
          "404" : {
            "description" : "Aucun conge n'existe  dans la BDD avec l'ID fourni"
          }
        }
      }
    },
...
    "/testCovid/v1/departement/{idDepartement}" : {
      "get" : {
        "tags" : [ "departement" ],
        "summary" : "rechercher un departement par ID",
        "description" : "Cette methode permet de rechercher  un departement par son ID ",
        "operationId" : "findById",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "idDepartement",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "la departement a ete trouver dans la BDD",
            "schema" : {
              "$ref" : "#/definitions/DepartementDTO"
            }
          },
          "404" : {
            "description" : "Aucun departement n'existe  dans la BDD avec l'ID fourni"
          }
        }
      }
    },
...
    "/testCovid/v1/detailResultat/{idDetailResultat}" : {
      "get" : {
        "tags" : [ "detailResultat" ],
        "summary" : "rechercher un detail Resultat par ID",
        "description" : "Cette methode permet de rechercher  un detail Resultat par son ID ",
        "operationId" : "findById",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "idDetailResultat",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "la detail Resultat a ete trouver dans la BDD",
            "schema" : {
              "$ref" : "#/definitions/DetailResultatDTO"
            }
          },
          "404" : {
            "description" : "Aucun detail Resultat n'existe  dans la BDD avec l'ID fourni"
          }
        }
      }
    },
...
    "/testCovid/v1/employe/{idEmploye}" : {
      "get" : {
        "tags" : [ "testCovidv1employe" ],
        "summary" : "rechercher un employe par ID",
        "description" : "Cette methode permet de rechercher  un employe par son ID ",
        "operationId" : "findById",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "idEmploye",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "la employe a ete trouver dans la BDD",
            "schema" : {
              "$ref" : "#/definitions/EmployeDTO"
            }
          },
          "404" : {
            "description" : "Aucun employe n'existe  dans la BDD avec l'ID fourni"
          }
        }
      }
    },
...
    "/testCovid/v1/facture/{idFacture}" : {
      "get" : {
        "tags" : [ "facture" ],
        "summary" : "rechercher d'une facture par ID",
        "description" : "Cette methode permet de rechercher d'une facture par son ID ",
        "operationId" : "findById",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "idFacture",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "la facture a ete trouver dans la BDD",
            "schema" : {
              "$ref" : "#/definitions/FactureDTO"
            }
          },
          "404" : {
            "description" : "Aucun facture n'existe  dans la BDD avec l'ID fourni"
          }
        }
      }
    },
...
    "/testCovid/v1/fournisseur/{idFournisseur}" : {
      "get" : {
        "tags" : [ "fournisseur" ],
        "summary" : "rechercher d'une fournisseur par ID",
        "description" : "Cette methode permet de rechercher d'une fournisseur par son ID ",
        "operationId" : "findById",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "idFournisseur",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "la fournisseur a ete trouver dans la BDD",
            "schema" : {
              "$ref" : "#/definitions/FournisseurDTO"
            }
          },
          "404" : {
            "description" : "Aucun fournisseur n'existe  dans la BDD avec l'ID fourni"
          }
        }
      }
    },
...
    "/testCovid/v1/labo/{idLabo}" : {
      "get" : {
        "tags" : [ "testCovidv1labo" ],
        "summary" : "rechercher un labo par ID",
        "description" : "Cette methode permet de rechercher  un labo par son ID ",
        "operationId" : "findById",
        "produces" : [ "application/json" ],
        "responses" : {
          "200" : {
            "description" : " labo a ete trouver dans la BDD",
            "schema" : {
              "$ref" : "#/definitions/LaboDTO"
            }
          },
          "404" : {
            "description" : "Aucun labo n'existe  dans la BDD avec l'ID fourni"
          }
        }
      }
    },
...
    "/testCovid/v1/mouvement/{idMouvement}" : {
      "get" : {
        "tags" : [ "testCovidv1mouvement" ],
        "summary" : "rechercher un mouvement par ID",
        "description" : "Cette methode permet de rechercher  un mouvement par son ID ",
        "operationId" : "findById",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "idMouvement",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "la patient a ete trouver dans la BDD",
            "schema" : {
              "$ref" : "#/definitions/MouvementDTO"
            }
          },
          "404" : {
            "description" : "Aucun patient n'existe  dans la BDD avec l'ID fourni"
          }
        }
      }
    },
...
    "/testCovid/v1/operation/{idOperation}" : {
      "get" : {
        "tags" : [ "testCovidv1operation" ],
        "summary" : "rechercher un operation par ID",
        "description" : "Cette methode permet de rechercher  un operation par son ID ",
        "operationId" : "findById",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "idOperation",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "la operation a ete trouver dans la BDD",
            "schema" : {
              "$ref" : "#/definitions/OperationDTO"
            }
          },
          "404" : {
            "description" : "Aucun operation n'existe  dans la BDD avec l'ID fourni"
          }
        }
      }
    },
...
    "/testCovid/v1/patient/{idPatient}" : {
      "get" : {
        "tags" : [ "testCovidv1patient" ],
        "summary" : "rechercher un patient par ID",
        "description" : "Cette methode permet de rechercher  un patient par son ID ",
        "operationId" : "findById",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "idPatient",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "la patient a ete trouver dans la BDD",
            "schema" : {
              "$ref" : "#/definitions/PatientDTO"
            }
          },
          "404" : {
            "description" : "Aucun patient n'existe  dans la BDD avec l'ID fourni"
          }
        }
      }
    },
...
    "/testCovid/v1/preleveur/{idPreleveur}" : {
      "get" : {
        "tags" : [ "preleveur" ],
        "summary" : "rechercher un preleveur par ID",
        "description" : "Cette methode permet de rechercher  un preleveur par son ID ",
        "operationId" : "findById",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "idPreleveur",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "la preleveur a ete trouver dans la BDD",
            "schema" : {
              "$ref" : "#/definitions/PreleveurDTO"
            }
          },
          "404" : {
            "description" : "Aucun preleveur n'existe  dans la BDD avec l'ID fourni"
          }
        }
      }
    },
...
    "/testCovid/v1/reservation/{idReservation}" : {
      "get" : {
        "tags" : [ "reservation" ],
        "summary" : "rechercher un reservation par ID",
        "description" : "Cette methode permet de rechercher  un reservation par son ID ",
        "operationId" : "findById",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "idReservation",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "la reservation a ete trouver dans la BDD",
            "schema" : {
              "$ref" : "#/definitions/ReservationDTO"
            }
          },
          "404" : {
            "description" : "Aucun reservation n'existe  dans la BDD avec l'ID fourni"
          }
        }
      }
    },
...
    "/testCovid/v1/role/{idRole}" : {
      "get" : {
        "tags" : [ "testCovidv1roles" ],
        "summary" : "rechercher un role par ID",
        "description" : "Cette methode permet de rechercher  un role par son ID ",
        "operationId" : "findById",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "idRole",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "la patient a ete trouver dans la BDD",
            "schema" : {
              "$ref" : "#/definitions/RolesDTO"
            }
          },
          "404" : {
            "description" : "Aucun patient n'existe  dans la BDD avec l'ID fourni"
          }
        }
      }
    },
...
    "/testCovid/v1/stock/filter{dateStock}" : {
      "get" : {
        "tags" : [ "stock" ],
        "summary" : "rechercher un stock par date de Stock",
        "description" : "Cette methode permet de rechercher un stock par date de Stock  ",
        "operationId" : "findBydateStock",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "dateStock",
          "in" : "path",
          "required" : true,
          "type" : "string",
          "format" : "date-time"
        } ],
        "responses" : {
          "200" : {
            "description" : "la stock a ete trouver dans la BDD",
            "schema" : {
              "$ref" : "#/definitions/StockDTO"
            }
          },
          "404" : {
            "description" : "Aucun stock n'existe  dans la BDD avec le CODE fourni"
          }
        }
      }
    },
    "/testCovid/v1/stock/{idStock}" : {
      "get" : {
        "tags" : [ "stock" ],
        "summary" : "rechercher un stock par ID",
        "description" : "Cette methode permet de rechercher  un stock par son ID ",
        "operationId" : "findById",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "idStock",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "la stock a ete trouver dans la BDD",
            "schema" : {
              "$ref" : "#/definitions/StockDTO"
            }
          },
          "404" : {
            "description" : "Aucun stock n'existe  dans la BDD avec l'ID fourni"
          }
        }
      }
    },
...
    "/testCovid/v1/typeAnalyse/{idTypeAnalyse}" : {
      "get" : {
        "tags" : [ "typeAnalyse" ],
        "summary" : "rechercher un typeAnalyse par ID",
        "description" : "Cette methode permet de rechercher  un typeAnalyse par son ID ",
        "operationId" : "findById",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "idTypeAnalyse",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "la typeAnalyse a ete trouver dans la BDD",
            "schema" : {
              "$ref" : "#/definitions/TypeAnalyseDTO"
            }
          },
          "404" : {
            "description" : "Aucun typeAnalyse n'existe  dans la BDD avec l'ID fourni"
          }
        }
      }
    }
  },
  "definitions" : {
    "AbsenceDTO" : {
      "type" : "object",
      "properties" : {
        "idAbsence" : {
          "type" : "integer",
          "format" : "int32"
        },
        "dateDebut" : {
          "type" : "string",
          "format" : "date-time"
        },
        "dateFin" : {
          "type" : "string",
          "format" : "date-time"
        },
        "justifie" : {
          "type" : "string"
        },
        "motif" : {
          "type" : "string"
        },
        "pieceJustificatifString" : {
          "type" : "string"
        },
        "employe" : {
          "$ref" : "#/definitions/EmployeDTO"
        }
      }
    },
    "AnalyseMedicalDTO" : {
      "type" : "object",
      "properties" : {
        "idAnalyse" : {
          "type" : "integer",
          "format" : "int32"
        },
        "dateAnalyse" : {
          "type" : "string",
          "format" : "date-time"
        },
        "descriprtion" : {
          "type" : "string"
        },
        "prixAnalyse" : {
          "type" : "number"
        },
        "resultat" : {
          "type" : "boolean"
        },
        "typeAnalyse" : {
          "$ref" : "#/definitions/TypeAnalyseDTO"
        },
        "reservation" : {
          "$ref" : "#/definitions/ReservationDTO"
        }
      }
    },
    "ArticleDTO" : {
      "type" : "object",
      "properties" : {
        "idArticle" : {
          "type" : "integer",
          "format" : "int32"
        },
        "prixArticle" : {
          "type" : "string"
        },
        "typeArticle" : {
          "type" : "string"
        },
        "codeBarre" : {
          "type" : "string"
        },
        "nomArticle" : {
          "type" : "string"
        }
      }
    },
...
    "BilanFinancierDTO" : {
      "type" : "object",
      "properties" : {
        "idBilan" : {
          "type" : "integer",
          "format" : "int32"
        },
        "description" : {
          "type" : "string"
        }
      }
    },
    "BiologisteDTO" : {
      "type" : "object",
      "properties" : {
        "idBiologiste" : {
          "type" : "integer",
          "format" : "int32"
        },
        "nom" : {
          "type" : "string"
        },
        "prenom" : {
          "type" : "string"
        },
        "email" : {
          "type" : "string"
        },
        "sexe" : {
          "type" : "string"
        },
        "telephone" : {
          "type" : "string"
        },
        "codeBiologiste" : {
          "type" : "string"
        },
        "service" : {
          "type" : "string"
        },
        "departement" : {
          "$ref" : "#/definitions/DepartementDTO"
        }
      }
    },
    "BonCommandeDTO" : {
      "type" : "object",
      "properties" : {
        "idBonCommande" : {
          "type" : "integer",
          "format" : "int32"
        },
        "dateCommande" : {
          "type" : "string",
          "format" : "date-time"
        },
        "fournisseur" : {
          "$ref" : "#/definitions/FournisseurDTO"
        },
        "article" : {
          "$ref" : "#/definitions/ArticleDTO"
        },
        "quantiteCommande" : {
          "type" : "integer",
          "format" : "int32"
        }
      }
    },
    "CompteDTO" : {
      "type" : "object",
      "properties" : {
        "idCompte" : {
          "type" : "integer",
          "format" : "int32"
        },
        "libCompte" : {
          "type" : "string"
        },
        "numeroCompte" : {
          "type" : "string"
        }
      }
    },
    "CongeDTO" : {
      "type" : "object",
      "properties" : {
        "idConge" : {
          "type" : "integer",
          "format" : "int32"
        },
        "dateDebut" : {
          "type" : "string",
          "format" : "date-time"
        },
        "dateFin" : {
          "type" : "string",
          "format" : "date-time"
        },
        "employe" : {
          "$ref" : "#/definitions/EmployeDTO"
        }
      }
    },
    "DepartementDTO" : {
      "type" : "object",
      "properties" : {
        "idDepartement" : {
          "type" : "integer",
          "format" : "int32"
        },
        "nomDepartement" : {
          "type" : "string"
        }
      }
    },
    "DetailResultatDTO" : {
      "type" : "object",
      "properties" : {
        "idResultat" : {
          "type" : "integer",
          "format" : "int32"
        },
        "description" : {
          "type" : "string"
        },
        "valeur" : {
          "type" : "number"
        },
        "biologiste" : {
          "$ref" : "#/definitions/BiologisteDTO"
        },
        "analyseMedical" : {
          "$ref" : "#/definitions/AnalyseMedicalDTO"
        }
      }
    },
    "EmployeDTO" : {
      "type" : "object",
      "properties" : {
        "id_employe" : {
          "type" : "integer",
          "format" : "int32"
        },
        "nom" : {
          "type" : "string"
        },
        "prenom" : {
          "type" : "string"
        },
        "email" : {
          "type" : "string"
        },
        "cin" : {
          "type" : "string"
        },
        "adresse" : {
          "type" : "string"
        },
        "telephone" : {
          "type" : "string"
        },
        "salaire" : {
          "type" : "number",
          "format" : "double"
        },
        "password" : {
          "type" : "string"
        },
        "date_embauchDate" : {
          "type" : "string",
          "format" : "date-time"
        },
        "photo" : {
          "type" : "string"
        },
        "labo" : {
          "$ref" : "#/definitions/LaboDTO"
        }
      }
    },
    "FactureDTO" : {
      "type" : "object",
      "properties" : {
        "idFacture" : {
          "type" : "integer",
          "format" : "int32"
        },
        "dateFacture" : {
          "type" : "string",
          "format" : "date-time"
        },
        "montant" : {
          "type" : "number"
        },
        "analyseMedical" : {
          "$ref" : "#/definitions/AnalyseMedicalDTO"
        }
      }
    },
    "FournisseurDTO" : {
      "type" : "object",
      "properties" : {
        "idFournisseur" : {
          "type" : "integer",
          "format" : "int32"
        },
        "nom" : {
          "type" : "string"
        },
        "prenom" : {
          "type" : "string"
        },
        "telephone" : {
          "type" : "string"
        },
        "adresse" : {
          "type" : "string"
        }
      }
    },
    "LaboDTO" : {
      "type" : "object",
      "properties" : {
        "idLabo" : {
          "type" : "integer",
          "format" : "int32"
        },
        "nomLabo" : {
          "type" : "string"
        },
        "codeFiscal" : {
          "type" : "string"
        },
        "numTel" : {
          "type" : "string"
        },
        "email" : {
          "type" : "string"
        },
        "siteWeb" : {
          "type" : "string"
        }
      }
    },
    "MouvementDTO" : {
      "type" : "object",
      "properties" : {
        "idMouvement" : {
          "type" : "integer",
          "format" : "int32"
        },
        "typeMouvement" : {
          "type" : "string"
        },
        "detailMouvement" : {
          "type" : "string"
        },
        "frais" : {
          "type" : "number",
          "format" : "double"
        },
        "employe" : {
          "$ref" : "#/definitions/EmployeDTO"
        }
      }
    },
    "OperationDTO" : {
      "type" : "object",
      "properties" : {
        "idOperation" : {
          "type" : "integer",
          "format" : "int32"
        },
        "dateOperation" : {
          "type" : "string",
          "format" : "date-time"
        },
        "typeOperation" : {
          "type" : "string"
        },
        "description" : {
          "type" : "string"
        },
        "compte" : {
          "$ref" : "#/definitions/CompteDTO"
        },
        "annee" : {
          "type" : "integer",
          "format" : "int32"
        },
        "montant" : {
          "type" : "number",
          "format" : "float"
        }
      }
    },
    "PatientDTO" : {
      "type" : "object",
      "properties" : {
        "idPatient" : {
          "type" : "integer",
          "format" : "int32"
        },
        "nom" : {
          "type" : "string"
        },
        "prenom" : {
          "type" : "string"
        },
        "email" : {
          "type" : "string"
        },
        "sexe" : {
          "type" : "string"
        },
        "telephone" : {
          "type" : "string"
        },
        "codePatient" : {
          "type" : "string"
        },
        "cni" : {
          "type" : "string"
        },
        "dateNaissance" : {
          "type" : "string",
          "format" : "date-time"
        },
        "adressse" : {
          "type" : "string"
        }
      }
    },
    "PreleveurDTO" : {
      "type" : "object",
      "properties" : {
        "idPreleveur" : {
          "type" : "integer",
          "format" : "int32"
        },
        "nom" : {
          "type" : "string"
        },
        "prenom" : {
          "type" : "string"
        },
        "email" : {
          "type" : "string"
        },
        "sexe" : {
          "type" : "string"
        },
        "telephone" : {
          "type" : "string"
        },
        "specialite" : {
          "type" : "string"
        },
        "analyseMedical" : {
          "$ref" : "#/definitions/AnalyseMedicalDTO"
        }
      }
    },
...
    "ReservationDTO" : {
      "type" : "object",
      "properties" : {
        "idReservation" : {
          "type" : "integer",
          "format" : "int32"
        },
        "dateReservation" : {
          "type" : "string",
          "format" : "date-time"
        },
        "description" : {
          "type" : "string"
        },
        "patient" : {
          "$ref" : "#/definitions/PatientDTO"
        }
      }
    },
    "RolesDTO" : {
      "type" : "object",
      "properties" : {
        "idRole" : {
          "type" : "integer",
          "format" : "int32"
        },
        "roleName" : {
          "type" : "string"
        }
      }
    },
    "StockDTO" : {
      "type" : "object",
      "properties" : {
        "idStock" : {
          "type" : "integer",
          "format" : "int32"
        },
        "dateStock" : {
          "type" : "string",
          "format" : "date-time"
        },
        "article" : {
          "$ref" : "#/definitions/ArticleDTO"
        },
        "quatiteStock" : {
          "type" : "string"
        }
      }
    },
    "TypeAnalyseDTO" : {
      "type" : "object",
      "properties" : {
        "idTypeAnalyse" : {
          "type" : "integer",
          "format" : "int32"
        },
        "typeAnalyse" : {
          "type" : "string"
        }
      }
    }
  }
}